feaat(settings): add text fields for changing settings

+++ app/src/main/java/org/mromichov/guessthesong/presentation/settings/SettingsUiState.kt
@@ -2,8 +2,10 @@ package org.mromichov.guessthesong.presentation.settings
 
 sealed class SettingsUiState {
     object Loading : SettingsUiState()
+
+    // snippetLength - seconds
     data class Success(
         val roundsNumber: Int,
-        val snippetLength: Long
+        val snippetLength: Int
     ) : SettingsUiState()
 }
